Changed files to reflect use of new database table.
I used the free US Zipcode data from geonames.org to
create a table called zips_2015.  Then I updated genjson.php
to use the new table, and I updated index.php because the
data in the new table had the latitude values and not the
negative of the latitude values.